PWM Resolution now 0.1%
Changed the resolution of the PWM to be 0.1%. This means e.g. sending 437 => 43.7%

+++ Modules/SimpleKeyboard/SimpleKeyboard_User.h
@@ -17,7 +17,7 @@
 
 
 extern uint8_t keyPress;
-extern uint8_t blowerPWM;
+extern uint16_t blowerPWM;
 extern tau_pwm_instance_ctrl_t g_timer0_ctrl;
 
 #if (USE_KBD_ENUM)
